Still More Chapter 2 Gold Challenge Struggle (Minor Change

+++ textbook_work/robert_hieger/RandomItems/RandomItems/BNRContainer.h
@@ -40,7 +40,7 @@
 
 // Read Only getter
 
-- (float) totalValue;
+- (int) totalValue;
 
 // Custom initializer for BNRContainer:
