+++ textbook_work/robert_hieger/RandomItems/RandomItems/BNRContainer.m
@@ -34,7 +34,7 @@
     
 }   // end containerName getter
 
-- (float) totalValue {
+- (int) totalValue {
     
     return _totalValue;
